GM_VIP_Automation-original/requirements.txt
# Python dependencies for GM VIP Automation scripts
#
# validate_capl.py  – stdlib only, no extra packages needed
# merge_reports.py  – stdlib only, no extra packages needed
# email_report.py   – stdlib only (smtplib, ssl, email), no extra packages needed
# Serial.py         – requires pyserial for COM-port communication with the
#                     programmable power supply and UART capture
#
# Install on the Windows bench agent:
#   pip install -r requirements.txt

pyserial==3.5
